% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRpus.lang.en-package.R
\docType{package}
\name{koRpus.lang.en-package}
\alias{koRpus.lang.en}
\alias{koRpus.lang.en-package}
\title{\packageTitle{koRpus.lang.en}}
\description{
\packageDescription{koRpus.lang.en}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab koRpus.lang.en\cr
Type: \tab Package\cr
Version: \tab 0.1-4\cr
Date: \tab 2020-10-24\cr
Depends: \tab R (>= 3.1),koRpus (>= 0.11-2)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=koRpus\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=koRpus}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/koRpus.lang.en/issues}
}

}
\author{
\packageAuthor{koRpus.lang.en}

Maintainer: \packageMaintainer{koRpus.lang.en}
}
